Move AudioPlayer to kit module

+++ Recordings-OO-MVC/app/src/main/java/com/madappgang/recordings/activities/RecorderActivity.kt
@@ -197,7 +197,7 @@ class RecorderActivity :
 
     private fun updateTime() {
         val totalRecordingTime = recordingTime + audioRecorder.getProgress()
-        time.text = time.formatMilliseconds(totalRecordingTime)
+        time.text = time.formatMilliseconds(totalRecordingTime.toInt())
     }
 
     private fun applyNotStartedStateForButton() {
